% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_PD.R, R/check_PSD.R
\name{check_PD}
\alias{check_PD}
\title{Check Positive semi-definite
Checks if the given matrix is positive definite.}
\usage{
check_PD(M, tol = 1e-09)

check_PD(M, tol = 1e-09)
}
\arguments{
\item{M}{Matrix to check whether it is positive definite.}

\item{tol}{Tolerance when comparing smallest eigenvalue to 0.}
}
\value{
TRUE if the matrix is PSD, FALSE otherwise.

TRUE if the matrix is PSD, FALSE otherwise.
}
\description{
Check Positive semi-definite
Checks if the given matrix is positive definite.

Check Positive semi-definite
Checks if the given matrix is positive definite.
}
