% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_PSD.R
\name{check_PSD}
\alias{check_PSD}
\title{Check Positive semi-definite
Checks if the given matrix is positive semi-definite.}
\usage{
check_PSD(M, tol = 1e-09)
}
\arguments{
\item{M}{Matrix to check whether it is positive semi-definite.}

\item{tol}{Tolerance when comparing smallest eigenvalue to 0.}
}
\value{
TRUE if the matrix is PSD, FALSE otherwise.
}
\description{
Check Positive semi-definite
Checks if the given matrix is positive semi-definite.
}
