% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_threshold.R
\name{compute_threshold}
\alias{compute_threshold}
\title{Knockoff threshold}
\usage{
compute_threshold(W, q, plus)
}
\arguments{
\item{W}{the set of p test statistics.}

\item{q}{the desired FDR.}

\item{plus}{a boolean T/F, true for the Knockoff+ procedure and false for the Knockoff procedure.}
}
\value{
A threshold T to use for variable selection.
}
\description{
Calculates the threshold T for variable selection for the Knockoff and Knockoff+ procedures.
}
