% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knockoffs.R
\name{create_knockoffs}
\alias{create_knockoffs}
\title{Create knockoffs}
\source{
Barber and Candes,
Controlling the false discovery rate via knockoffs.
Ann. Statist. 43 (2015), no. 5, 2055--2085.
https://projecteuclid.org/euclid.aos/1438606853
}
\usage{
create_knockoffs(X, y, method = c("sdp", "equi"), randomize = FALSE)
}
\arguments{
\item{X}{n-by-p design matrix, with \eqn{n > p}.}

\item{y}{response vector of length n.
This is used in the setting in which \eqn{n < 2p}, to estimate sigma hat and generate additional rows in X and y.}

\item{method}{either "equi" or "sdp" (default: "sdp").
These are two different methods to generate the p-dimensional s vector, which is used to constuct the knockoff variables.}

\item{randomize}{If true, U tilde is a random matrix. If false, U tilde is the second half of the
matrix Q, where Q is part of the QR decomposition of an nx2p matrix \eqn{[U\ 0]}, where \eqn{X = UDV^T}.
default is false.}
}
\value{
A list containing:
 \item{X}{n-by-p design matrix, rescaled so that \eqn{||X_j||^2_2 = 1}, and augmented if \eqn{n < 2p}.}
 \item{X_knock}{n-by-p matrix of knockoff variables.}
 \item{y}{vector of observed responses (augmented if \eqn{n < 2p}). }
}
\description{
Creates knockoff variables for fixed design matrix.
}
\details{
This creates p knockoff variables for a fixed design matrix of size n by p. These knockoffs can be
used to provably control the false discovery rate (FDR) for the linear model with iid Gaussian errors.
}
\examples{
p <- 50; n <- 100; k <- 5
X <- matrix(stats::rnorm(n*p), nrow = n)
true_covar <- sample(p, k)
beta <- 5 * (1:p \%in\% true_covar)
y <- X \%*\% beta + stats::rnorm(n, mean = 0, sd = 1)
knock <- create_knockoffs(X, y, method = 'sdp')

}
