% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_permutations.R
\name{create_permutations}
\alias{create_permutations}
\title{Create permuted variables}
\source{
Barber and Candes,
Controlling the false discovery rate via knockoffs.
Ann. Statist. 43 (2015), no. 5, 2055--2085.
https://projecteuclid.org/euclid.aos/1438606853
}
\usage{
create_permutations(X, y)
}
\arguments{
\item{X}{n-by-p design matrix, with \eqn{n > p}.}

\item{y}{response vector of length n.
This is used in the setting in which \eqn{n < 2p}, to estimate sigma hat and generate additional rows in X and y.}
}
\value{
A list containing:
 \item{X}{n-by-p design matrix, rescaled so that \eqn{||X_j||^2_2 = 1}, and augmented if \eqn{n < 2p}.}
 \item{X_perm}{n-by-p matrix of permuted variables.}
 \item{y}{vector of observed responses (augmented if \eqn{n < 2p}). }
}
\description{
Creates a permuted version of the fixed design matrix.
}
\details{
This creates p permuted variables for a fixed design matrix of size n by p.
}
