% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sdp.R
\name{create_sdp}
\alias{create_sdp}
\title{SDP Knockoffs
Creates fixed-X knockoffs using the SDP method}
\usage{
create_sdp(X)
}
\arguments{
\item{X}{Design matrix to create knockoffs for.}
}
\value{
The knockoff nxp design matrix.
}
\description{
SDP Knockoffs
Creates fixed-X knockoffs using the SDP method
}
