% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stat_lambda_max.R
\name{get_stat_lambda_max}
\alias{get_stat_lambda_max}
\title{Lambda max test statistic
Creates W_j test statistics. Performs penalized regression for a range of lambda values on the augmented
X matrix \eqn{[X\ \widetilde{X}]}. Gets \eqn{Z_j = \sup\{\lambda:\hat{\beta}_j \neq 0\}} for each
original and knockoff variables. Computes the test statistic \eqn{W_j = max(Z_j, Z_{j+p)*-1\cdot I(Z_j < Z_{j+p})})}.}
\usage{
get_stat_lambda_max(X, Xk, y)
}
\arguments{
\item{X}{n x p X matrix.}

\item{Xk}{n x p knockoff matrix.}

\item{y}{n dimensional y vector.}
}
\value{
A vector of p test statistics \eqn{W_j}.
}
\description{
Lambda max test statistic
Creates W_j test statistics. Performs penalized regression for a range of lambda values on the augmented
X matrix \eqn{[X\ \widetilde{X}]}. Gets \eqn{Z_j = \sup\{\lambda:\hat{\beta}_j \neq 0\}} for each
original and knockoff variables. Computes the test statistic \eqn{W_j = max(Z_j, Z_{j+p)*-1\cdot I(Z_j < Z_{j+p})})}.
}
