% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.knockoffs.R
\name{prelim-knockoffs-package}
\alias{prelim-knockoffs-package}
\alias{prelim-knockoffs}
\title{Knockoff package documentation (implemented for prelim).}
\description{
The knockoff controls the false discovery rate in the Gaussian linear model with fixed-X.
}
\details{
This runs the knockoff filter for a given X matrix and y vector. This controls the false discovery
rate at a given level.
}
\references{
Barber and Candes,
Controlling the false discovery rate via knockoffs.
Ann. Statist. 43 (2015), no. 5, 2055--2085.
https://projecteuclid.org/euclid.aos/1438606853
}
