% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BHq.R
\name{run_BHq}
\alias{run_BHq}
\title{Run Benjamini-Hochberg procedure}
\usage{
run_BHq(X, y, fdr, method = c("base", "log", "white_noise"))
}
\arguments{
\item{X}{A nxp design matrix.}

\item{y}{A vector of n responses.}

\item{fdr}{The desired false discovery rate.}

\item{method}{either "base", "log", or "white_noise" (default: "base").}
}
\value{
A list containing:
 \item{X}{n-by-p design matrix.}
 \item{y}{vector of observed responses. }
 \item{Z}{vector of test statistics.}
 \item{thres}{data-dependent threshold.}
 \item{selected}{list of selected variables.}
}
\description{
Runs the either the original Benjamini-Hochberg procedure, the BH procedure with a log correction,
or the BH procedure with whitened noise.
}
