% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BHq_base.R
\name{run_BHq_base}
\alias{run_BHq_base}
\title{Run Benjamini-Hochberg procedure}
\usage{
run_BHq_base(X, y, fdr)
}
\arguments{
\item{X}{A nxp design matrix.}

\item{y}{A vector of n responses.}

\item{fdr}{The desired false discovery rate.}
}
\value{
A list containing:
 \item{X}{n-by-p design matrix.}
 \item{y}{vector of observed responses. }
 \item{Z}{vector of test statistics.}
 \item{thres}{data-dependent threshold.}
 \item{selected}{list of selected variables.}
}
\description{
Runs the original Benjamini-Hochberg procedure.
}
