% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_variables}
\alias{select_variables}
\title{Variable selection}
\usage{
select_variables(W, thresh)
}
\arguments{
\item{W}{the set of p test statistics.}

\item{thresh}{threshold for variable selection.}
}
\value{
A list of selected variables.
}
\description{
Selects variables using the W_j statistics and threshold provided.
}
