% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sdp.R
\name{solve_sdp}
\alias{solve_sdp}
\title{SDP using Rdsdp
Solves the optimation problem to choose s for the SDP knockoff method, using a SDP solver
implemented in Rdsdp. Specifically, this solves the problem:
\eqn{\text{Maximize}\sum_{j=1}^p s_j} subject to the linear constraint \eqn{0 \leq s_j \leq 1\ \forall\ s_j}
and the PSD constraint \eqn{2\Sigma - \text{diag}\{s\} \succeq 0}.}
\usage{
solve_sdp(Sigma, gaptol = 1e-06, maxit = 1000)
}
\arguments{
\item{Sigma}{Normalized empirical covariance matrix to optimize s for.}

\item{gaptol}{The tolerance for the duality gap as a fraction of the objective functions (a parameter in Rdsdp).}

\item{maxit}{The maximum number of iterations allowed (a paramter in Rdsdp).}
}
\value{
The vector of s values that solve the above optimization problem.
}
\description{
SDP using Rdsdp
Solves the optimation problem to choose s for the SDP knockoff method, using a SDP solver
implemented in Rdsdp. Specifically, this solves the problem:
\eqn{\text{Maximize}\sum_{j=1}^p s_j} subject to the linear constraint \eqn{0 \leq s_j \leq 1\ \forall\ s_j}
and the PSD constraint \eqn{2\Sigma - \text{diag}\{s\} \succeq 0}.
}
