% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_sign.R
\name{svd_sign}
\alias{svd_sign}
\title{SVD with fixed sign
Performs SVD and then adjusts the sign of each column of the U matrix so that the value
with the greatest magnitude is positive. It fixes the signs of the V matrix accordingly.}
\usage{
svd_sign(X)
}
\arguments{
\item{X}{Design matrix to create knockoffs for.}
}
\value{
The SVD of X: U, a vector of singular values d, V, such that \eqn{X = Udiag\{d\}V^T}.
}
\description{
SVD with fixed sign
Performs SVD and then adjusts the sign of each column of the U matrix so that the value
with the greatest magnitude is positive. It fixes the signs of the V matrix accordingly.
}
