% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_equi.R
\name{create_equi}
\alias{create_equi}
\title{Equi-correlated Knockoffs
Creates fixed-X knockoffs using the equi-correlated method}
\usage{
create_equi(X)
}
\arguments{
\item{X}{Design matrix to create knockoffs for.}
}
\value{
The knockoff nxp design matrix.
}
\description{
Equi-correlated Knockoffs
Creates fixed-X knockoffs using the equi-correlated method
}
