% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sdp.R
\name{create_sdp}
\alias{create_sdp}
\title{SDP Knockoffs
Creates fixed-X knockoffs using the SDP method}
\usage{
create_sdp(X, randomize = FALSE)
}
\arguments{
\item{X}{Design matrix to create knockoffs for.}

\item{randomize}{If true, U tilde is a random matrix. If false, U tilde is the second half of the
matrix Q, where Q is part of the QR decomposition of an nx2p matrix \eqn{[U\ 0]}, where \eqn{X = UDV^T}.
default is false.}
}
\value{
The knockoff nxp design matrix.
}
\description{
SDP Knockoffs
Creates fixed-X knockoffs using the SDP method
}
